requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
requests==2.32.5
urllib3==2.6.3
azure-functions
azure-storage-blob
flake8
black
